+++ src/gql/directives/Transform.php
@@ -52,7 +52,7 @@ class Transform extends Directive
                 DirectiveLocation::FIELD,
             ],
             'args' => TransformArguments::getArguments(),
-            'description' => 'This directive is used to return a URL for an [asset transform](https://craftcms.com/docs/3.x/image-transforms.html). It accepts the same arguments you would use for a transform in Craft and adds the `immediately` argument.',
+            'description' => 'This directive is used to return a URL for an [asset transform](https://craftcms.com/docs/3.x/image-transforms.html). It accepts the same arguments you would use for a transform in Craft.',
         ]));
     }
 
@@ -77,7 +77,6 @@ class Transform extends Directive
             return $value;
         }
 
-        $generateNow = $arguments['immediately'] ?? Craft::$app->getConfig()->general->generateTransformsBeforePageLoad;
         $transform = Gql::prepareTransformArguments($arguments);
 
         // If this directive is applied to an entire Asset
@@ -102,7 +101,7 @@ class Transform extends Directive
             case 'width':
                 return $source->getWidth($transform);
             case 'url':
-                return $source->getUrl($transform, $generateNow);
+                return $source->getUrl($transform);
         }
 
         return $value;